Time: 12 ms (15.77%), Space: 11.2 MB (77.73%) - LeetHub

+++ 2696-minimum-string-length-after-removing-substrings/2696-minimum-string-length-after-removing-substrings.cpp
@@ -10,6 +10,6 @@ public:
                 st.push(s[i]);
             }
         }
-        return st.size();
+        return (int)st.size();
     }
 };
